fix: replace uniform_int_distribution<uint8_t> with int+cast for MSVC compatibility MSVC rejects std::uniform_int_distribution<uint8_t> (it is undefined behavior per the C++ standard — the type parameter must be a short, int, long, long long, or their unsigned counterparts). Replace all three occurrences with uniform_int_distribution<int> and static_cast<uint8_t>(...), fixing Windows CI build errors in: - benchmarks/src/allocator_benchmark.cpp - benchmarks/custom/fragmentation_benchmark.cpp - benchmarks/custom/compression_drift_benchmark.cpp

+++ benchmarks/custom/compression_drift_benchmark.cpp
@@ -68,12 +68,12 @@ struct DriftParams {
 // ratio close to 1.0 = incompressible (random data)
 void generate_data_with_ratio(std::vector<uint8_t>& data, double compress_ratio, std::mt19937& gen) {
     std::uniform_real_distribution<double> prob(0.0, 1.0);
-    std::uniform_int_distribution<uint8_t> byte_dist(0, 255);
+    std::uniform_int_distribution<int> byte_dist(0, 255);
 
     for (auto& byte : data) {
         if (prob(gen) < compress_ratio) {
             // Random byte (incompressible)
-            byte = byte_dist(gen);
+            byte = static_cast<uint8_t>(byte_dist(gen));
         } else {
             // Zero (highly compressible)
             byte = 0;
